+++ ios/InteractiveReader/InteractiveReader/Features/InteractivePlayer/InteractivePlayerView+Layout.swift
@@ -58,7 +58,6 @@ extension InteractivePlayerView {
         trackpadSwipeLayer
         shortcutHelpOverlay
         keyboardShortcutLayer
-        swiftUIKeyboardShortcutLayer
     }
 
     @ViewBuilder